Allow using URLs to refer to NamedNodes
We will now assume that the most common entity that will be written
is a NamedNode, and that it's easier for the library consumer to be
able to simply pass around their IRI's. That means that if they
want to store literals, they will now have to explicitly wrap those
into a Literal, since we have to be able to discern between a
Node's IRI and a plain string.