Add pointers to the getAll* methods in the docs
It should be clear to library consumers that there might be more
values than the one they're fetching, which would get ignored using
the singular versions of the methods.
(I've also considered naming them `getOne*`, but that would get
rather unwieldy for such a common operation rather quickly. But who
knows, maybe some day, if this turns out to be a major source of
confusion still.)