Replace rdflib by N3.js
By replacing rdflib with N3, we get a far lighter-weight version
of Tripledoc. This greatly reduces our bundle size, and we will
later be able to take advantage of N3's store for better
performance.

Most of the code we lose is for features irrelevant to
Tripledoc, such as support for formats other than Turtle.
However, it also means that many of the workarounds included in
rdflib are now gone, so some specific things - such as TimBL's
WebID being served without the needed headeres - will no longer
work.