Add support for reading local Subjects
Although we don't support _writing_ Blank Nodes (yet?), others
might have already written them to the Pod, so we should support
reading them.
In this commit, I've attempted to represent them using new
instances of TripleSubjects, that can be queried for their own
values. This appears to mesh well with the idea that Blank Nodes
are mostly used for more complex data structures that represent
specific properties of a Subject, rather than an entity by itself
that needs to be passed around.