Remove references to the Document in the Store
They are redundant anyway, since each Document gets its own Store.
But primarily, this is in preparation for allowing Documents to
lazily initialise their URL, e.g. when they're created in a
Container, in which case their URL will only be known _after_
saving. When a Document's URL is not known yet, we should still be
able to query its Store, even if it's empty still.