Allow creating Documents in Containers
This feature required significant rework, that mostly came down to
adding two "intermediate" types representing TripleDocuments in two
different stages in their lifetime: the BareTripleDocument, for
Documents that do not exist yet on the Pod and also do not have a
known location yet (i.e. those that will be created in a
Container), and the InMemoryTripleDocument, that does not exist on
the Pod yet, but for which we know where we want to create it.

The main takeaway from this is that this will result in certain
methods not being available yet when they are moot at that point in
a Document's lifecycle, such as `.asRef` for a Document that does
not have a known location yet. Before, they would simply return
`null`.