Append `WithRef` to LocalTripleDocument
This should help with claryfing that it's "on an equal footing"
with the LocalTripleDocumentForContainer, rather than the latter
being some specific instance of the former. Also, it helpfully
comes later in the alphabet now and is hence listed below
LocalTripleDocumentForContainer, which somewhat nicely aligns with
it having some additional methods.